Adds handling of AIR invoke event
The invoke event is triggered when the application starts
and also if the user tries to start a second instance of the app.

The invoke handler is used to pull out a filename from the command-line arguments
and open the appropriate database file. This means that if Tarpo is set
as the default application for .tarpo files, double-clicking on a .tarpo file
will open that file (even if Tarpo is already running - intelligently prompting
the user if they want to close the current file)